Added authorization requirement for put, delete endpoints in /posts | Created search, limit, skip functions on the get endpoint in /posts

+++ app/routers/post.py
@@ -1,11 +1,11 @@
 from typing import List
-from ..schemas import Post, PostResponse
+from ..schemas import Post, PostResponse, TokenSchema
 from sqlalchemy.orm import Session
 from fastapi import Response, status, HTTPException, Depends, APIRouter
 
 from ..database import get_db
 from .. import models
-
+from .. import oauth2
 
 router = APIRouter(
     prefix="/posts",
@@ -17,8 +17,11 @@ router = APIRouter(
 
 
 @router.get("/", response_model=List[PostResponse])
-def get_all_posts(db: Session = Depends(get_db)):
-    posts = db.query(models.Post).all()
+def get_all_posts(db: Session = Depends(get_db), limit: int = 5, skip: int = 0, search: str | None = ""):
+    # posts = db.query(models.Post).all()
+    # posts = db.query(models.Post).limit(limit).all()
+    posts = db.query(models.Post).filter(
+        models.Post.title.contains(search)).limit(limit).offset(skip).all()
     return posts
 
 
@@ -39,8 +42,10 @@ def get_single_post(id: int, response: Response, db: Session = Depends(get_db)):
 
 
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model=PostResponse)
-def create_post(payload: Post, db: Session = Depends(get_db)):
-    new_post = models.Post(**payload.dict())
+def create_post(payload: Post, db: Session = Depends(get_db), token_user_data: dict = Depends(oauth2.get_current_user)):
+    new_post = models.Post(
+        **payload.dict(), owner=token_user_data.user_id)
+    print(token_user_data, "This is what you stored in the token")
     db.add(new_post)
     db.commit()
     db.refresh(new_post)
@@ -51,12 +56,15 @@ def create_post(payload: Post, db: Session = Depends(get_db)):
 
 
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete(id: int, db: Session = Depends(get_db)):
+def delete(id: int, db: Session = Depends(get_db), token_user_data: dict = Depends(oauth2.get_current_user)):
     post_query = db.query(models.Post).filter(models.Post.id == id)
     post = post_query.first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail="This post doesn't exist")
+    if post.owner != token_user_data.user_id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN, detail="You are not allowed to perform this action.")
     post_query.delete(synchronize_session=False)
     db.commit()
     return None
@@ -66,12 +74,15 @@ def delete(id: int, db: Session = Depends(get_db)):
 
 
 @router.put("/{id}", response_model=PostResponse)
-def update(id: int, payload: Post, db: Session = Depends(get_db)):
+def update(id: int, payload: Post, db: Session = Depends(get_db), token_user_data: dict = Depends(oauth2.get_current_user)):
     post_query = db.query(models.Post).filter(models.Post.id == id)
     post = post_query.first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail="This post doesn't exist!")
+    if post.owner != token_user_data.user_id:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED, detail="You are not allowed to perform this action.")
     post_query.update(payload.dict(), synchronize_session=False)
     db.commit()
     db.refresh(post)
